+++ R/export.sc_hplm.R
@@ -9,6 +9,7 @@ export.sc_hplm <- function(object, caption = NA, footnote = NA, filename = NA,
                            kable_options = list(), 
                            round = 2,
                            nice = TRUE,
+                           casewise = FALSE,
                            ...) {
   
   kable_options <- .join_kabel(kable_options)
@@ -21,16 +22,6 @@ export.sc_hplm <- function(object, caption = NA, footnote = NA, filename = NA,
     )
   }
   
-  summary_model <- summary(object$hplm)
-  if (object$model$ICC) {
-    ICC <- sprintf(
-      "<i>ICC</i> = %.3f, <i>L</i> = %.1f, <i>p</i> = %.3f", 
-      object$ICC$value, object$ICC$L, object$ICC$p
-    )
-  } else {
-    ICC <- ""
-  }
-  
   footnote <- c(
     paste0("Estimation method ", object$model$estimation.method),
     paste0("Slope estimation method: ", object$model$interaction.method),
@@ -38,18 +29,27 @@ export.sc_hplm <- function(object, caption = NA, footnote = NA, filename = NA,
     paste0(object$N, " cases")
   )
   
+  if (casewise) {
+    out <- .export_casewise(
+      object, caption, footnote, filename ,
+      kable_styling_options, 
+      kable_options, 
+      round,
+      ...
+    )
+    return(out)
+  }
+  
+  summary_model <- summary(object$hplm)
+  
   out <- as.data.frame(summary(object$hplm)$tTable)
-  
   row.names(out) <- .plm.row.names(row.names(out), object)
-  
   colnames(out) <- c("B", "SE", "df", "t", "p")
   
   md <- data.frame(
     "SD" = round(as.numeric(VarCorr(object$hplm)[, "StdDev"]), 3)
   )
-  rownames(md) <- names(VarCorr(object$hplm)[, 2])
-  
-  row.names(md) <- .plm.row.names(row.names(md), object)
+  rownames(md) <- .plm.row.names(names(VarCorr(object$hplm)[, 2]), object)
   
   if (object$model$lr.test) {
     if (is.null(object$LR.test[[1]]$L.Ratio)) {
@@ -131,4 +131,24 @@ export.sc_hplm <- function(object, caption = NA, footnote = NA, filename = NA,
   
   table
   
+}
+
+.export_casewise <- function(object, caption = NA, footnote = NA, filename = NA,
+                             kable_styling_options = list(), 
+                             kable_options = list(), 
+                             round = 2,
+                             ...) {
+  
+  out <- coef(object, casewise = TRUE)
+  
+  table <- .create_table(
+    out, 
+    kable_options, 
+    kable_styling_options, 
+    caption = caption,
+    footnote = footnote
+  )
+  
+  if (!is.na(filename)) .save_export(table, filename)
+  table
 }