+++ NEWS.md
@@ -13,6 +13,9 @@
 - `rand_test()`: New option for `statistic`: `SMD` calculates the standardized mean difference as Hedge's g with Durlak correction. `W-test` computes Wilcoxon tests and compares average W statistics. `T-test` computes T-tests and compares average t-Values. `NAP` and `NAP decreasing` for Non-overlap of all pairs.
 - `nap()`: added Cohen's d and R-Squared effects.
 - `export()`: `select` argument for `nap`; new for `pem()`, `pet()`
+- `coef.sc_hplm()`: new `casewise` argument. If set TRUE, returns the effect estimatinons casewise.
+- `print.sc_hplm()`: new `casewise` argument. If set TRUE, returns the effect estimatinons casewise.
+- `export.sc_hplm()`: new `casewise` argument. If set TRUE, returns the effect estimatinons casewise.
 
 ## New examples